Move country specific tests into folders and run separately
