+++ v1-to-v2-data-migration/tests/unit/birthResolver.test.ts
@@ -3,7 +3,8 @@ import { assertEquals } from 'https://deno.land/std@0.210.0/assert/mod.ts'
 import {
   buildBirthResolver,
   buildBirthEventRegistration,
-} from '../utils/test-helpers.ts'
+} from '../utils/testHelpers.ts'
+import { buildPhoneNumber, stripCountryCode } from '../utils/phoneBuilder.ts'
 
 // Construct birthResolver as in migrate.ipynb
 const birthResolver = buildBirthResolver()
@@ -77,63 +78,6 @@ Deno.test('birthResolver - child fields', async (t) => {
     const declareAction = result.actions.find((a) => a.type === 'DECLARE')
 
     assertEquals(declareAction?.declaration['child.birthLocation'], 'facility1')
-  })
-
-  await t.step(
-    'should resolve child.birthLocation.privateHome for PRIVATE_HOME',
-    () => {
-      const registration = buildBirthEventRegistration({
-        eventLocation: {
-          type: 'PRIVATE_HOME',
-          address: {
-            line: ['123', 'Main St', 'City', '', '', 'URBAN'],
-            district: 'District1',
-            state: 'State1',
-            city: 'City1',
-            country: 'FAR',
-            postalCode: '12345',
-          },
-        },
-      })
-
-      const result = transform(registration, birthResolver, 'birth')
-      const declareAction = result.actions.find((a) => a.type === 'DECLARE')
-
-      assertEquals(
-        declareAction?.declaration['child.birthLocation.privateHome']
-          ?.administrativeArea,
-        'District1'
-      )
-      assertEquals(
-        declareAction?.declaration['child.birthLocation.privateHome']?.country,
-        'FAR'
-      )
-    }
-  )
-
-  await t.step('should resolve child.birthLocation.other for OTHER', () => {
-    const registration = buildBirthEventRegistration({
-      eventLocation: {
-        type: 'OTHER',
-        address: {
-          line: ['456', 'Other St', 'Town'],
-          district: 'District2',
-          state: 'State2',
-          city: 'City2',
-          country: 'FAR',
-          postalCode: '54321',
-        },
-      },
-    })
-
-    const result = transform(registration, birthResolver, 'birth')
-    const declareAction = result.actions.find((a) => a.type === 'DECLARE')
-
-    assertEquals(
-      declareAction?.declaration['child.birthLocation.other']
-        ?.administrativeArea,
-      'District2'
-    )
   })
 
   await t.step('should resolve child.attendantAtBirth', () => {
@@ -360,33 +304,6 @@ Deno.test('birthResolver - mother fields', async (t) => {
     assertEquals(declareAction?.declaration['mother.previousBirths'], 2)
   })
 
-  await t.step('should resolve mother.address', () => {
-    const registration = buildBirthEventRegistration({
-      mother: {
-        address: [
-          {
-            type: 'PRIMARY_ADDRESS',
-            line: ['10', 'Mother St', 'City', '', '', 'URBAN'],
-            country: 'FAR',
-            district: 'District1',
-            state: 'State1',
-            city: 'City1',
-            postalCode: '11111',
-          },
-        ],
-      },
-    })
-
-    const result = transform(registration, birthResolver, 'birth')
-    const declareAction = result.actions.find((a) => a.type === 'DECLARE')
-
-    assertEquals(declareAction?.declaration['mother.address']?.country, 'FAR')
-    assertEquals(
-      declareAction?.declaration['mother.address']?.administrativeArea,
-      'District1'
-    )
-  })
-
   await t.step('should resolve mother.brn', () => {
     const registration = buildBirthEventRegistration({
       mother: {
@@ -588,91 +505,6 @@ Deno.test('birthResolver - father fields', async (t) => {
     assertEquals(declareAction?.declaration['father.occupation'], 'Engineer')
   })
 
-  await t.step('should resolve father.address', () => {
-    const registration = buildBirthEventRegistration({
-      father: {
-        address: [
-          {
-            type: 'PRIMARY_ADDRESS',
-            line: ['20', 'Father Ave', 'Town'],
-            country: 'FAR',
-            district: 'District2',
-            state: 'State2',
-            city: 'City2',
-            postalCode: '22222',
-          },
-        ],
-      },
-    })
-
-    const result = transform(registration, birthResolver, 'birth')
-    const declareAction = result.actions.find((a) => a.type === 'DECLARE')
-
-    assertEquals(declareAction?.declaration['father.address']?.country, 'FAR')
-    assertEquals(
-      declareAction?.declaration['father.address']?.administrativeArea,
-      'District2'
-    )
-  })
-
-  await t.step(
-    'should resolve father.addressSameAs when addresses match',
-    () => {
-      const sameAddress = {
-        type: 'PRIMARY_ADDRESS' as const,
-        line: ['10', 'Same St', 'City'],
-        country: 'FAR',
-        district: 'District1',
-        state: 'State1',
-        city: 'City1',
-        postalCode: '11111',
-      }
-
-      const registration = buildBirthEventRegistration({
-        mother: { address: [sameAddress] },
-        father: { address: [sameAddress] },
-      })
-
-      const result = transform(registration, birthResolver, 'birth')
-      const declareAction = result.actions.find((a) => a.type === 'DECLARE')
-
-      assertEquals(declareAction?.declaration['father.addressSameAs'], 'YES')
-    }
-  )
-
-  await t.step(
-    'should resolve father.addressSameAs when addresses differ',
-    () => {
-      const registration = buildBirthEventRegistration({
-        mother: {
-          address: [
-            {
-              type: 'PRIMARY_ADDRESS',
-              line: ['10', 'Mother St'],
-              country: 'FAR',
-              district: 'District1',
-            },
-          ],
-        },
-        father: {
-          address: [
-            {
-              type: 'PRIMARY_ADDRESS',
-              line: ['20', 'Father Ave'],
-              country: 'FAR',
-              district: 'District2',
-            },
-          ],
-        },
-      })
-
-      const result = transform(registration, birthResolver, 'birth')
-      const declareAction = result.actions.find((a) => a.type === 'DECLARE')
-
-      assertEquals(declareAction?.declaration['father.addressSameAs'], 'NO')
-    }
-  )
-
   await t.step('should resolve father.brn', () => {
     const registration = buildBirthEventRegistration({
       father: {
@@ -775,48 +607,20 @@ Deno.test('birthResolver - informant fields', async (t) => {
   })
 
   await t.step(
-    'should resolve informant.address for non-special informant',
-    () => {
-      const registration = buildBirthEventRegistration({
-        informant: {
-          relationship: 'GRANDFATHER',
-          address: [
-            {
-              type: 'PRIMARY_ADDRESS',
-              line: ['30', 'Informant Rd'],
-              country: 'FAR',
-              district: 'District3',
-              state: 'State3',
-              city: 'City3',
-              postalCode: '33333',
-            },
-          ],
-        },
-      })
-
-      const result = transform(registration, birthResolver, 'birth')
-      const declareAction = result.actions.find((a) => a.type === 'DECLARE')
-
-      assertEquals(
-        declareAction?.declaration['informant.address']?.administrativeArea,
-        'District3'
-      )
-    }
-  )
-
-  await t.step(
     'should resolve informant.phoneNo with country code stripped',
     () => {
+      const fullPhoneNumber = buildPhoneNumber('0987654321')
       const registration = buildBirthEventRegistration({
         registration: {
           trackingId: 'B123456',
-          contactPhoneNumber: '+260987654321',
+          contactPhoneNumber: fullPhoneNumber,
         },
       })
 
       const result = transform(registration, birthResolver, 'birth')
       const declareAction = result.actions.find((a) => a.type === 'DECLARE')
 
+      // Should strip country code and return local number with leading zero
       assertEquals(
         declareAction?.declaration['informant.phoneNo'],
         '0987654321'
